route logic and organize network configs
