+++ wormhole-connect/src/config/devnet/rpcs.ts
@@ -1,0 +1,10 @@
+export const DEVNET_RPC_MAPPING = {
+  ethereum: process.env.REACT_APP_ETHEREUM_DEVNET_RPC,
+  osmosis: process.env.REACT_APP_OSMOSIS_DEVNET_RPC,
+  wormchain: process.env.REACT_APP_WORMCHAIN_DEVNET_RPC,
+  terra2: process.env.REACT_APP_TERRA2_DEVNET_RPC,
+};
+
+export const DEVNET_REST_MAPPING = {
+  sei: process.env.REACT_APP_SEI_REST,
+};
