+++ wormhole-connect/src/config/mainnet/rpcs.ts
@@ -1,0 +1,20 @@
+export const MAINNET_RPC_MAPPING = {
+  ethereum: process.env.REACT_APP_ETHEREUM_RPC,
+  solana: process.env.REACT_APP_SOLANA_RPC,
+  polygon: process.env.REACT_APP_POLYGON_RPC,
+  bsc: process.env.REACT_APP_BSC_RPC,
+  avalanche: process.env.REACT_APP_AVALANCHE_RPC,
+  fantom: process.env.REACT_APP_FANTOM_RPC,
+  celo: process.env.REACT_APP_CELO_RPC,
+  moonbeam: process.env.REACT_APP_MOONBEAM_RPC,
+  sui: process.env.REACT_APP_SUI_RPC,
+  aptos: process.env.REACT_APP_APTOS_RPC,
+  sei: process.env.REACT_APP_SEI_RPC,
+  base: process.env.REACT_APP_BASE_RPC,
+  osmosis: process.env.REACT_APP_OSMOSIS_RPC,
+  wormchain: process.env.REACT_APP_WORMCHAIN_RPC,
+};
+
+export const MAINNET_REST_MAPPING = {
+  sei: process.env.REACT_APP_SEI_REST,
+};
