+++ wormhole-connect/src/config/mainnet/tokens.ts
@@ -1,179 +1,6 @@
-import { CONFIG } from '@wormhole-foundation/wormhole-connect-sdk';
-import { NetworksConfig, TokenConfig, GasEstimates } from './types';
+import { Icon, TokensConfig } from '../types';
 
-const { chains } = CONFIG.MAINNET;
-
-export enum Icon {
-  'AVAX' = 1,
-  'BNB',
-  'BSC',
-  'CELO',
-  'ETH',
-  'FANTOM',
-  'POLYGON',
-  'SOLANA',
-  'USDC',
-  'GLMR',
-  'DAI',
-  'USDT',
-  'BUSD',
-  'WBTC',
-  'SUI',
-  'APT',
-  'SEI',
-  'BASE',
-  'OSMO',
-  'BONK',
-  'TBTC',
-  'WSTETH',
-}
-
-export const MAINNET_NETWORKS: NetworksConfig = {
-  ethereum: {
-    ...chains.ethereum!,
-    displayName: 'Ethereum',
-    explorerUrl: 'https://etherscan.io/',
-    explorerName: 'Etherscan',
-    gasToken: 'ETH',
-    chainId: 1,
-    icon: Icon.ETH,
-    automaticRelayer: true,
-    maxBlockSearch: 2000,
-  },
-  bsc: {
-    ...chains.bsc!,
-    displayName: 'BSC',
-    explorerUrl: 'https://bscscan.com/',
-    explorerName: 'BscScan',
-    gasToken: 'BNB',
-    chainId: 56,
-    icon: Icon.BSC,
-    automaticRelayer: true,
-    maxBlockSearch: 2000,
-  },
-  polygon: {
-    ...chains.polygon!,
-    displayName: 'Polygon',
-    explorerUrl: 'https://polygonscan.com/',
-    explorerName: 'PolygonScan',
-    gasToken: 'MATIC',
-    chainId: 137,
-    icon: Icon.POLYGON,
-    automaticRelayer: true,
-    maxBlockSearch: 1000,
-  },
-  avalanche: {
-    ...chains.avalanche!,
-    displayName: 'Avalanche',
-    explorerUrl: 'https://snowtrace.io/',
-    explorerName: 'Snowtrace',
-    gasToken: 'WAVAX',
-    chainId: 43114,
-    icon: Icon.AVAX,
-    automaticRelayer: true,
-    maxBlockSearch: 2000,
-  },
-  fantom: {
-    ...chains.fantom!,
-    displayName: 'Fantom',
-    explorerUrl: 'https://ftmscan.com/',
-    explorerName: 'FTMscan',
-    gasToken: 'FTM',
-    chainId: 250,
-    icon: Icon.FANTOM,
-    automaticRelayer: true,
-    maxBlockSearch: 2000,
-  },
-  celo: {
-    ...chains.celo!,
-    displayName: 'Celo',
-    explorerUrl: 'https://explorer.celo.org/mainnet/',
-    explorerName: 'Celo Explorer',
-    gasToken: 'CELO',
-    chainId: 42220,
-    icon: Icon.CELO,
-    automaticRelayer: true,
-    maxBlockSearch: 2000,
-  },
-  moonbeam: {
-    ...chains.moonbeam!,
-    displayName: 'Moonbeam',
-    explorerUrl: 'https://moonscan.io/',
-    explorerName: 'Moonscan',
-    gasToken: 'GLMR',
-    chainId: 1284,
-    icon: Icon.GLMR,
-    automaticRelayer: true,
-    maxBlockSearch: 2000,
-  },
-  solana: {
-    ...chains.solana!,
-    displayName: 'Solana',
-    explorerUrl: 'https://explorer.solana.com/',
-    explorerName: 'Solana Explorer',
-    gasToken: 'SOL',
-    chainId: 0,
-    icon: Icon.SOLANA,
-    automaticRelayer: false,
-    maxBlockSearch: 2000,
-  },
-  sui: {
-    ...chains.sui!,
-    displayName: 'Sui',
-    explorerUrl: 'https://explorer.sui.io/',
-    explorerName: 'Sui Explorer',
-    gasToken: 'SUI',
-    chainId: 0,
-    icon: Icon.SUI,
-    automaticRelayer: true,
-    maxBlockSearch: 0,
-  },
-  aptos: {
-    ...chains.aptos!,
-    displayName: 'Aptos',
-    explorerUrl: 'https://explorer.aptoslabs.com/',
-    explorerName: 'Aptos Explorer',
-    gasToken: 'APT',
-    chainId: 0,
-    icon: Icon.APT,
-    automaticRelayer: false,
-    maxBlockSearch: 0,
-  },
-  base: {
-    ...chains.base!,
-    displayName: 'Base',
-    explorerUrl: 'https://basescan.org/',
-    explorerName: 'BaseScan',
-    gasToken: 'ETHbase',
-    chainId: 8453,
-    icon: Icon.BASE,
-    maxBlockSearch: 2000,
-  },
-  osmosis: {
-    ...chains.osmosis!,
-    displayName: 'Osmosis',
-    explorerUrl: 'https://mintscan.io/osmosis/',
-    explorerName: 'MintScan',
-    gasToken: 'OSMO',
-    chainId: 'osmosis-1',
-    icon: Icon.OSMO,
-    automaticRelayer: false,
-    maxBlockSearch: 0,
-  },
-  wormchain: {
-    ...chains.wormchain!,
-    displayName: 'Wormchain',
-    explorerUrl: '',
-    explorerName: '',
-    gasToken: 'WORM',
-    chainId: '',
-    icon: Icon.OSMO,
-    automaticRelayer: false,
-    maxBlockSearch: 0,
-  },
-};
-
-export const MAINNET_TOKENS: { [key: string]: TokenConfig } = {
+export const MAINNET_TOKENS: TokensConfig = {
   ETH: {
     key: 'ETH',
     symbol: 'ETH',
@@ -656,98 +483,5 @@ export const MAINNET_TOKENS: { [key: string]: TokenConfig } = {
     decimals: {
       default: 5,
     },
-  },
-};
-
-export const MAINNET_GAS_ESTIMATES: GasEstimates = {
-  ethereum: {
-    sendNative: 100000,
-    sendToken: 150000,
-    sendNativeWithRelay: 200000,
-    sendTokenWithRelay: 300000,
-    sendCCTPWithRelay: 300000,
-    sendCCTPManual: 150000,
-    claim: 300000,
-  },
-  polygon: {
-    sendNative: 100000,
-    sendToken: 150000,
-    sendNativeWithRelay: 200000,
-    sendTokenWithRelay: 250000,
-    claim: 300000,
-  },
-  bsc: {
-    sendNative: 100000,
-    sendToken: 200000,
-    sendNativeWithRelay: 200000,
-    sendTokenWithRelay: 300000,
-    claim: 300000,
-  },
-  avalanche: {
-    sendNative: 100000,
-    sendToken: 150000,
-    sendNativeWithRelay: 200000,
-    sendTokenWithRelay: 300000,
-    sendCCTPWithRelay: 300000,
-    sendCCTPManual: 150000,
-    claim: 300000,
-  },
-  fantom: {
-    sendNative: 100000,
-    sendToken: 150000,
-    sendNativeWithRelay: 250000,
-    sendTokenWithRelay: 300000,
-    claim: 300000,
-  },
-  celo: {
-    sendNative: 150000,
-    sendToken: 150000,
-    sendNativeWithRelay: 300000,
-    sendTokenWithRelay: 300000,
-    claim: 300000,
-  },
-  moonbeam: {
-    sendNative: 100000,
-    sendToken: 150000,
-    sendNativeWithRelay: 200000,
-    sendTokenWithRelay: 300000,
-    claim: 300000,
-  },
-  solana: {
-    sendNative: 15000,
-    sendToken: 15000,
-    claim: 25000,
-  },
-  sui: {
-    sendNative: 5000000,
-    sendToken: 5000000,
-    sendNativeWithRelay: 5000000,
-    sendTokenWithRelay: 5000000,
-    claim: 200000000,
-  },
-  aptos: {
-    sendNative: 40,
-    sendToken: 40,
-    claim: 1250,
-  },
-  sei: {
-    claim: 1000000,
-    sendToken: 1000000,
-    sendNative: 1000000,
-  },
-  base: {
-    claim: 1000000,
-    sendNative: 1000000,
-    sendToken: 1000000,
-  },
-  wormchain: {
-    sendNative: 0,
-    sendToken: 0,
-    claim: 0,
-  },
-  osmosis: {
-    sendNative: 0,
-    sendToken: 0,
-    claim: 0,
   },
 };
