+++ wormhole-connect/src/config/testnet/rpcs.ts
@@ -1,0 +1,20 @@
+export const TESTNET_RPC_MAPPING = {
+  goerli: process.env.REACT_APP_GOERLI_RPC,
+  mumbai: process.env.REACT_APP_MUMBAI_RPC,
+  bsc: process.env.REACT_APP_BSC_TESTNET_RPC,
+  fuji: process.env.REACT_APP_FUJI_RPC,
+  fantom: process.env.REACT_APP_FANTOM_TESTNET_RPC,
+  alfajores: process.env.REACT_APP_ALFAJORES_RPC,
+  solana: process.env.REACT_APP_SOLANA_DEVNET_RPC,
+  moonbasealpha: process.env.REACT_APP_MOONBASE_RPC,
+  sui: process.env.REACT_APP_SUI_TESTNET_RPC,
+  aptos: process.env.REACT_APP_APTOS_TESTNET_RPC,
+  sei: process.env.REACT_APP_SEI_TESTNET_RPC,
+  basegoerli: process.env.REACT_APP_BASE_GOERLI_RPC,
+  osmosis: process.env.REACT_APP_OSMOSIS_TESTNET_RPC,
+  wormchain: process.env.REACT_APP_WORMCHAIN_TESTNET_RPC,
+};
+
+export const TESTNET_REST_MAPPING = {
+  sei: process.env.REACT_APP_SEI_REST,
+};
