+++ wormhole-connect/src/config/testnet/tokens.ts
@@ -1,188 +1,6 @@
-import { CONFIG } from '@wormhole-foundation/wormhole-connect-sdk';
-import { NetworksConfig, TokenConfig, GasEstimates } from './types';
+import { Icon, TokensConfig } from '../types';
 
-const { chains } = CONFIG.TESTNET;
-export enum Icon {
-  'AVAX' = 1,
-  'BNB',
-  'BSC',
-  'CELO',
-  'ETH',
-  'FANTOM',
-  'POLYGON',
-  'SOLANA',
-  'USDC',
-  'GLMR',
-  'DAI',
-  'USDT',
-  'BUSD',
-  'WBTC',
-  'SUI',
-  'APT',
-  'SEI',
-  'BASE',
-  'OSMO',
-  'BONK',
-  'TBTC',
-  'WSTETH',
-}
-
-export const TESTNET_NETWORKS: NetworksConfig = {
-  goerli: {
-    ...chains.goerli!,
-    displayName: 'Goerli',
-    explorerUrl: 'https://goerli.etherscan.io/',
-    explorerName: 'Etherscan',
-    gasToken: 'ETH',
-    chainId: 5,
-    icon: Icon.ETH,
-    automaticRelayer: true,
-    maxBlockSearch: 2000,
-  },
-  mumbai: {
-    ...chains.mumbai!,
-    displayName: 'Mumbai',
-    explorerUrl: 'https://mumbai.polygonscan.com/',
-    explorerName: 'Polygonscan',
-    gasToken: 'MATIC',
-    chainId: 80001,
-    icon: Icon.POLYGON,
-    automaticRelayer: true,
-    maxBlockSearch: 1000,
-  },
-  bsc: {
-    ...chains.bsc!,
-    displayName: 'BSC',
-    explorerUrl: 'https://testnet.bscscan.com/',
-    explorerName: 'BscScan',
-    gasToken: 'BNB',
-    chainId: 97,
-    icon: Icon.BSC,
-    automaticRelayer: true,
-    maxBlockSearch: 2000,
-  },
-  fuji: {
-    ...chains.fuji!,
-    displayName: 'Fuji',
-    explorerUrl: 'https://testnet.snowtrace.io/',
-    explorerName: 'Snowtrace',
-    gasToken: 'AVAX',
-    chainId: 43113,
-    icon: Icon.AVAX,
-    automaticRelayer: true,
-    maxBlockSearch: 2000,
-  },
-  fantom: {
-    ...chains.fantom!,
-    displayName: 'Fantom',
-    explorerUrl: 'https://testnet.ftmscan.com/',
-    explorerName: 'FtmScan',
-    gasToken: 'FTM',
-    chainId: 4002,
-    icon: Icon.FANTOM,
-    automaticRelayer: true,
-    maxBlockSearch: 2000,
-  },
-  alfajores: {
-    ...chains.alfajores!,
-    displayName: 'Alfajores',
-    explorerUrl: 'https://explorer.celo.org/alfajores/',
-    explorerName: 'Celo Explorer',
-    gasToken: 'CELO',
-    chainId: 44787,
-    icon: Icon.CELO,
-    automaticRelayer: true,
-    maxBlockSearch: 2000,
-  },
-  moonbasealpha: {
-    ...chains.moonbasealpha!,
-    displayName: 'Moonbase',
-    explorerUrl: 'https://moonbase.moonscan.io/',
-    explorerName: 'Moonscan',
-    gasToken: 'GLMR',
-    chainId: 1287,
-    icon: Icon.GLMR,
-    automaticRelayer: true,
-    maxBlockSearch: 2000,
-  },
-  solana: {
-    ...chains.solana!,
-    displayName: 'Solana',
-    explorerUrl: 'https://explorer.solana.com/',
-    explorerName: 'Solana Explorer',
-    gasToken: 'SOL',
-    chainId: 0,
-    icon: Icon.SOLANA,
-    automaticRelayer: false,
-    maxBlockSearch: 2000,
-  },
-  sui: {
-    ...chains.sui!,
-    displayName: 'Sui',
-    explorerUrl: 'https://explorer.sui.io/',
-    explorerName: 'Sui Explorer',
-    gasToken: 'SUI',
-    chainId: 0,
-    icon: Icon.SUI,
-    automaticRelayer: true,
-    maxBlockSearch: 0,
-  },
-  aptos: {
-    ...chains.aptos!,
-    displayName: 'Aptos',
-    explorerUrl: 'https://explorer.aptoslabs.com/',
-    explorerName: 'Aptos Explorer',
-    gasToken: 'APT',
-    chainId: 0,
-    icon: Icon.APT,
-    maxBlockSearch: 0,
-  },
-  basegoerli: {
-    ...chains.basegoerli!,
-    displayName: 'Base Goerli',
-    explorerUrl: 'https://goerli.basescan.org/',
-    explorerName: 'BaseScan',
-    gasToken: 'ETHbase',
-    chainId: 84531,
-    icon: Icon.BASE,
-    maxBlockSearch: 2000,
-  },
-  sei: {
-    ...chains.sei!,
-    displayName: 'Sei',
-    explorerUrl: 'https://sei.explorers.guru/',
-    explorerName: 'Sei Explorer',
-    gasToken: 'SEI',
-    chainId: 0,
-    icon: Icon.SEI,
-    automaticRelayer: false,
-    maxBlockSearch: 0,
-  },
-  osmosis: {
-    ...chains.osmosis!,
-    displayName: 'Osmosis',
-    explorerUrl: 'https://testnet.mintscan.io/osmosis-testnet/',
-    explorerName: 'MintScan',
-    gasToken: 'OSMO',
-    chainId: 'osmo-test-5',
-    icon: Icon.OSMO,
-    automaticRelayer: false,
-    maxBlockSearch: 0,
-  },
-  wormchain: {
-    ...chains.wormchain!,
-    displayName: 'Wormchain',
-    explorerUrl: '',
-    explorerName: '',
-    gasToken: 'WORM',
-    chainId: 'wormchain-testnet-0',
-    icon: Icon.OSMO,
-    automaticRelayer: false,
-    maxBlockSearch: 0,
-  },
-};
-
-export const TESTNET_TOKENS: { [key: string]: TokenConfig } = {
+export const TESTNET_TOKENS: TokensConfig = {
   ETH: {
     key: 'ETH',
     symbol: 'ETH',
@@ -633,98 +451,5 @@ export const TESTNET_TOKENS: { [key: string]: TokenConfig } = {
     decimals: {
       default: 6,
     },
-  },
-};
-
-export const TESTNET_GAS_ESTIMATES: GasEstimates = {
-  goerli: {
-    sendNative: 100000,
-    sendToken: 150000,
-    sendNativeWithRelay: 200000,
-    sendTokenWithRelay: 300000,
-    sendCCTPWithRelay: 300000,
-    sendCCTPManual: 150000,
-    claim: 200000,
-  },
-  mumbai: {
-    sendNative: 200000,
-    sendToken: 150000,
-    sendNativeWithRelay: 200000,
-    sendTokenWithRelay: 250000,
-    claim: 200000,
-  },
-  bsc: {
-    sendNative: 100000,
-    sendToken: 200000,
-    sendNativeWithRelay: 200000,
-    sendTokenWithRelay: 300000,
-    claim: 175000,
-  },
-  fuji: {
-    sendNative: 100000,
-    sendToken: 150000,
-    sendNativeWithRelay: 200000,
-    sendTokenWithRelay: 300000,
-    sendCCTPWithRelay: 300000,
-    sendCCTPManual: 150000,
-    claim: 200000,
-  },
-  fantom: {
-    sendNative: 150000,
-    sendToken: 150000,
-    sendNativeWithRelay: 200000,
-    sendTokenWithRelay: 300000,
-    claim: 200000,
-  },
-  alfajores: {
-    sendNative: 100000,
-    sendToken: 100000,
-    sendNativeWithRelay: 300000,
-    sendTokenWithRelay: 300000,
-    claim: 175000,
-  },
-  moonbasealpha: {
-    sendNative: 100000,
-    sendToken: 200000,
-    sendNativeWithRelay: 200000,
-    sendTokenWithRelay: 300000,
-    claim: 200000,
-  },
-  solana: {
-    sendNative: 15000,
-    sendToken: 15000,
-    claim: 25000,
-  },
-  sui: {
-    sendNative: 20000000,
-    sendToken: 20000000,
-    sendNativeWithRelay: 20000000,
-    sendTokenWithRelay: 20000000,
-    claim: 20000000,
-  },
-  aptos: {
-    sendNative: 34,
-    sendToken: 34,
-    claim: 615,
-  },
-  sei: {
-    claim: 1000000,
-    sendNative: 1000000,
-    sendToken: 1000000,
-  },
-  basegoerli: {
-    claim: 100000,
-    sendNative: 1000000,
-    sendToken: 1000000,
-  },
-  wormchain: {
-    sendNative: 0,
-    sendToken: 0,
-    claim: 0,
-  },
-  osmosis: {
-    sendNative: 0,
-    sendToken: 0,
-    claim: 0,
   },
 };
