+++ wormhole-connect/src/utils/routes/cctpRelay.ts
@@ -8,8 +8,8 @@ import {
 } from '@wormhole-foundation/wormhole-connect-sdk';
 import { BigNumber, utils } from 'ethers';
 
-import { CHAINS, TOKENS } from 'config';
-import { TokenConfig, Route, PayloadType } from 'config/types';
+import { CHAINS, ROUTES, TOKENS } from 'config';
+import { TokenConfig, Route } from 'config/types';
 import {
   MAX_DECIMALS,
   getTokenDecimals,
@@ -18,7 +18,13 @@ import {
   getWrappedTokenId,
   getTokenById,
 } from 'utils';
-import { ParsedMessage, ParsedRelayerMessage, toChainId, wh } from 'utils/sdk';
+import {
+  ParsedMessage,
+  ParsedRelayerMessage,
+  PayloadType,
+  toChainId,
+  wh,
+} from 'utils/sdk';
 import { TransferWallet, signAndSendTransaction } from 'utils/wallet';
 import { NO_INPUT } from 'utils/style';
 import {
@@ -147,7 +153,9 @@ export class CCTPRelayRoute extends CCTPManualRoute {
     sourceChain: ChainName | ChainId,
     destChain: ChainName | ChainId,
   ): Promise<boolean> {
-    return false;
+    if (!(Route.CCTPManual in ROUTES)) {
+      return false;
+    }
 
     const sourceTokenConfig = TOKENS[sourceToken];
     const destTokenConfig = TOKENS[destToken];
@@ -400,7 +408,7 @@ export class CCTPRelayRoute extends CCTPManualRoute {
     if (!receipt) throw new Error(`No receipt for ${tx} on ${chain}`);
 
     const relayInfo = (await wh.getMessage(tx, chain)) as RelayTransferMessage;
-    if (relayInfo.payloadID !== PayloadType.AUTOMATIC)
+    if (relayInfo.payloadID !== PayloadType.Automatic)
       throw new Error('Transfer is not a relay transfer');
 
     // Get the CCTP log
@@ -433,7 +441,7 @@ export class CCTPRelayRoute extends CCTPManualRoute {
       sendTx: receipt.transactionHash,
       sender: receipt.from,
       amount: parsedCCTPLog.args.amount.toString(),
-      payloadID: PayloadType.MANUAL,
+      payloadID: PayloadType.Manual,
       recipient: recipient,
       toChain: getChainNameCCTP(parsedCCTPLog.args.destinationDomain),
       fromChain: fromChain,
