+++ wormhole-connect/src/utils/routes/cosmosGateway.ts
@@ -33,7 +33,7 @@ import { BigNumber, utils } from 'ethers';
 import { arrayify, base58, hexlify } from 'ethers/lib/utils.js';
 import { toChainId, wh } from 'utils/sdk';
 import { isCosmWasmChain } from 'utils/cosmos';
-import { CHAINS, CONFIG, TOKENS } from 'config';
+import { CHAINS, CONFIG, ROUTES, TOKENS } from 'config';
 import { Route } from 'config/types';
 import { MAX_DECIMALS, getTokenDecimals, toNormalizedDecimals } from '..';
 import { toDecimals, toFixedDecimals } from '../balance';
@@ -97,6 +97,10 @@ export class CosmosGatewayRoute extends BaseRoute {
     sourceChain: ChainName | ChainId,
     destChain: ChainName | ChainId,
   ): Promise<boolean> {
+    if (!(Route.CCTPManual in ROUTES)) {
+      return false;
+    }
+
     return (
       isCosmWasmChain(wh.toChainId(sourceChain)) ||
       isCosmWasmChain(wh.toChainId(destChain))
