+++ wormhole-connect/src/utils/routes/operator.ts
@@ -6,10 +6,11 @@ import {
 } from '@wormhole-foundation/wormhole-connect-sdk';
 import { BigNumber } from 'ethers';
 
+import { ROUTES } from 'config';
 import { TokenConfig, Route } from 'config/types';
 import { BridgeRoute } from './bridge';
 import { RelayRoute } from './relay';
-import { HashflowRoute } from './hashflow';
+// import { HashflowRoute } from './hashflow';
 import { CCTPRelayRoute } from './cctpRelay';
 import { CosmosGatewayRoute } from './cosmosGateway';
 import { ParsedMessage, PayloadType, getMessage, wh } from '../sdk';
@@ -26,15 +27,6 @@ import {
   CCTP_LOG_TokenMessenger_DepositForBurn,
 } from './cctpManual';
 
-// TODO: need to make this configurable
-export const listOfRoutes = [
-  Route.Bridge,
-  /*
-  Route.CCTPManual,
-  Route.CCTPManual,*/
-  Route.Relay,
-  Route.CosmosGateway,
-];
 export default class Operator {
   getRoute(route: Route): RouteAbstract {
     switch (route) {
@@ -50,9 +42,9 @@ export default class Operator {
       case Route.CCTPRelay: {
         return new CCTPRelayRoute();
       }
-      case Route.Hashflow: {
-        return new HashflowRoute();
-      }
+      // case Route.Hashflow: {
+      //   return new HashflowRoute();
+      // }
       case Route.CosmosGateway: {
         return new CosmosGatewayRoute();
       }
@@ -118,7 +110,7 @@ export default class Operator {
       return Route.CosmosGateway;
     }
 
-    return (message as ParsedMessage).payloadID === PayloadType.AUTOMATIC
+    return (message as ParsedMessage).payloadID === PayloadType.Automatic
       ? Route.Relay
       : Route.Bridge;
   }
@@ -131,6 +123,10 @@ export default class Operator {
     sourceChain: ChainName | ChainId,
     destChain: ChainName | ChainId,
   ): Promise<boolean> {
+    if (!(route in ROUTES)) {
+      return false;
+    }
+
     const r = this.getRoute(route);
     return await r.isRouteAvailable(
       sourceToken,
