+++ wormhole-connect/src/utils/sdk.ts
@@ -13,8 +13,8 @@ import { getWrappedTokenId } from '.';
 import { ENV, TOKENS, WH_CONFIG } from '../config';
 
 export enum PayloadType {
-  MANUAL = 1,
-  AUTOMATIC = 3,
+  Manual = 1,
+  Automatic = 3,
 }
 
 export const wh: WormholeContext = new WormholeContext(
