+++ wormhole-connect/src/views/Redeem/SendTo.tsx
@@ -4,7 +4,6 @@ import React, { useCallback, useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
 import { CHAINS } from 'config';
-import { Route } from 'config/types';
 import { RootState } from 'store';
 import { setRedeemTx, setTransferComplete } from 'store/redeem';
 import { displayAddress } from 'utils';
@@ -15,6 +14,7 @@ import {
   registerWalletSigner,
   switchNetwork,
 } from 'utils/wallet';
+import { PayloadType } from 'utils/sdk';
 
 import AlertBanner from 'components/AlertBanner';
 import Button from 'components/Button';
@@ -66,7 +66,7 @@ function SendTo() {
   }, [redeemTx, signedMessage, dispatch]);
 
   useEffect(() => {
-    if (!txData) return;
+    if (!txData || !routeType) return;
     const populate = async () => {
       let receiveTx: string | undefined;
       try {
@@ -91,6 +91,9 @@ function SendTo() {
   const claim = async () => {
     setInProgress(true);
     setClaimError('');
+    if (!routeType) {
+      throw new Error('Unknown route, cannot claim');
+    }
     if (!wallet || !isConnected) {
       setClaimError('Connect to receiving wallet');
       throw new Error('Connect to receiving wallet');
@@ -108,13 +111,13 @@ function SendTo() {
         registerWalletSigner(txData.toChain, TransferWallet.RECEIVING);
         await switchNetwork(networkConfig.chainId, TransferWallet.RECEIVING);
       }
-      if (!messageInfo) {
+      if (!signedMessage) {
         throw new Error('failed to get vaa, cannot redeem');
       }
       const txId = await new Operator().redeem(
         routeType,
         txData.toChain,
-        signedMessage!,
+        signedMessage,
         wallet.address,
       );
       dispatch(setRedeemTx(txId));
@@ -129,11 +132,11 @@ function SendTo() {
   };
 
   const loading =
-    txData.payloadID === Route.Bridge
+    txData.payloadID === PayloadType.Manual
       ? inProgress && !transferComplete
       : !transferComplete;
   const manualClaimText =
-    transferComplete || txData.payloadID === Route.Relay // todo: should be the other enum, should be named better than payload id
+    transferComplete || txData.payloadID === PayloadType.Automatic // todo: should be the other enum, should be named better than payload id
       ? ''
       : claimError
       ? 'Error please retry . . .'
@@ -152,7 +155,7 @@ function SendTo() {
       </InputContainer>
 
       {/* Claim button for manual transfers */}
-      {txData.payloadID === Route.Bridge && !transferComplete && (
+      {txData.payloadID === PayloadType.Manual && !transferComplete && (
         <>
           <Spacer height={8} />
           <AlertBanner
